+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Internal
 
 - Fix inconsistent overdue section header UI
+- Use `OverdueListSectionStates` to update overdue section states in `OverdueModel`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291